More refactor to accommodate more clients
